Refactor import logic to make missing license files recoverable.
Remove logic that tries to automatically create a patch for missing licenses.

Reorganize flow so that we do everything we can before reporting
a missing license file error, and leave the partially imported crate
in a usable state, so we can fix the problem and run `regenerate`.

Tested as follows:

1. Import khronos_api.

./crate_tool import khronos_api 3.1.0

Observe error message and create patch for license file.

./crate_tool regenerate khronos_api

2. Re-import spdx.

This crate was chosen because it has no patches and an empty cargo_embargo.json.

Remove spdx from pseudo_crate/Cargo.toml

Remove crates/spdx

./crate_tool import spdx 0.10.8

I observed with `git diff` that the only difference is last_upgrade_date in the METADATA file, which is expected.

Bug: 395132270
Test: Described in detail in commit message.
Change-Id: I570ca9508d87a304f37e196237e82a28251fbc6a