Keep track of license terms we don't need to follow.
When evaluating SPDX license expressions, keep track of both the
licenses we must follow, and those we do not need to follow.

For example, in the case of "Apache-2.0 OR MIT", we choose to follow
Apache-2.0, and so do not need to follow the MIT license.

This will allow us, when analyzing license files, to differentiate
between:
* License files we need to follow and keep around.
* License files mentioned in Cargo.toml that we don't need to follow and
  can therefore delete.
* License files for licenses not mentioned in Cargo.toml, which are
  unexpected and indicate a potential licensing and compliance problem
  requiring human intervention. For example, particularly in the case of
  *-sys crates that provide an interface to a C library, the Cargo.toml
  may contain the terms for the Rust code, but not the underlying C
  library that ships with the crate, which could have different,
  unacceptable terms.

Test: cargo test
Change-Id: Id1a5b856f8c125a57207b45f5c248febbbca8a01