Update the parser to parse validations
Summary: A simple parser can be used to dump the build graph in a machine-readable format, such as JSON or Python pickle. The parser was added in 2018 and hasn't been updated since.

Ninja 1.11 introduced validations that add a new token `|@` to describe the list of validations that are needed to be run.

This update is modifying the parser to parse build files generated by a build system.

Test: Successful Build on master branch

Change-Id: Ie06b5aa6ddb0e65f5e2a00ea939863dc2e054e02
Signed-off-by: Pratik Kathalkar <pkathalk@meta.com>